Store timing history and refactor helpers
